将部分int变量改为size_t

+++ robin/src/lexer/lexer.h
@@ -14,9 +14,9 @@ namespace Lexer {
         std::string source;
         std::vector<Token::Token> tokens;
 
-        int start = 0;
-        int current = 0;
-        int line = 1;
+        size_t start = 0;
+        size_t current = 0;
+        size_t line = 1;
         static std::map<std::string, Token::TokenType> keywords;
 
     public:
