+++ robin/src/token/token.cpp
@@ -61,7 +61,7 @@ namespace Token {
     };
 
 
-    Token::Token(TokenType type, std::string lexeme, std::any literal, int line)
+    Token::Token(TokenType type, std::string lexeme, std::any literal, size_t line)
             : type{type}, lexeme{std::move(lexeme)}, literal{std::move(literal)}, line{line} {}
