+++ robin/src/token/token.h
@@ -72,10 +72,10 @@ namespace Token {
         TokenType type;//type: 表示标记的类型，是一个枚举类型 TokenType 的实例。
         std::string lexeme;//lexeme: 表示标记在源代码中的文本形式，即标记的字符串表示。
         Object literal;//literal: 表示标记的字面量值，可以是任意类型的对象，用于表示标记的具体值。
-        int line;//line: 表示标记在源代码中所在的行号。
+        size_t line;//line: 表示标记在源代码中所在的行号。
 
     public:
-        Token(TokenType type, std::string lexeme, Object literal, int line);
+        Token(TokenType type, std::string lexeme, Object literal,size_t line);
 
         std::string toString() const;
